instalado pacotes e criados primeiros components
