+++ desafio-react-redux/src/Components/Content.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import Login from './Login';
+
+const Content = () => {
+  return (
+    <div>
+      <Login />
+    </div>
+  );
+};
+
+export default Content;
