+++ desafio-react-redux/src/Components/Header.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const Header = () => {
+  return (
+    <div>
+      <h1>Mini Dogs</h1>
+    </div>
+  );
+};
+
+export default Header;
