+++ desafio-react-redux/src/store/configureStore.js
@@ -3,14 +3,12 @@ import {
   configureStore,
   getDefaultMiddleware,
 } from '@reduxjs/toolkit';
-import contador from './contador';
-import modal from './modal';
+
 import login from './login';
-// import logger from './middleware/logger';logger
 import localStorage from './middleware/localStorage';
 
 const middleware = [...getDefaultMiddleware(), localStorage];
-const reducer = combineReducers({ contador, modal, login });
+const reducer = combineReducers({ login });
 
 const store = configureStore({ reducer, middleware });
 export default store;
